Move `LastUpdatedAt` to `core/data/time` package
